+++ server/src/permissions.py
@@ -13,7 +13,7 @@ def run_cmds(*cmds):
         cmd_name, exitcode, err = cmds[index], e.returncode, e.output
         print("ERROR RUNNING: '{}' OUTPUT: {}".format(cmd_name, out))
 
-def check_permissions(sudoPassword):
+def GUI_check_permissions(sudoPassword):
   # TODO: check for other operating systems
   print('Checking permissions...')
   ls_permissions = config.OSX_LS_BPF_PERMISSIONS
@@ -27,7 +27,7 @@ def check_permissions(sudoPassword):
   else:
     print('Permissions already set')
 
-def set_permissions(sudoPassword):
+def GUI_set_permissions(sudoPassword):
   print('Setting permissions, you may need to enter your password...')
   add_permissions = config.OSX_ADD_BPF_PERMISSIONS
   p = os.system('echo %s|sudo -S %s' % (sudoPassword, add_permissions))
@@ -36,7 +36,7 @@ def set_permissions(sudoPassword):
   run_cmds(enable_forwarding)
   print('Permissions set')
 
-def restore_permissions(sudoPassword):
+def GUI_restore_permissions(sudoPassword):
   print(' Restoring to default permissions...')
   subtract_permissions = config.OSX_SUBTR_BPF_PERMISSIONS
   p = os.system('echo %s|sudo -S %s' % (sudoPassword, subtract_permissions))
@@ -46,5 +46,33 @@ def restore_permissions(sudoPassword):
   print('Permissions restored')
 
 
+def check_permissions():
+  # TODO: check for other operating systems
+  print('Checking permissions...')
+  ls_permissions = config.OSX_LS_BPF_PERMISSIONS
+  permissions = subprocess.run(ls_permissions, shell=True, check=True, stdout=subprocess.PIPE)
+  permissions_str = permissions.stdout.decode()
+  if config.OSX_BPF_PERMISSIONS in permissions_str:
+    try:
+      set_permissions()
+    except:
+      print('Setting permissions failed')
+  else:
+    print('Permissions already set')
 
+def set_permissions():
+  print('Setting permissions, you may need to enter your password...')
+  add_permissions = config.OSX_ADD_BPF_PERMISSIONS
+  enable_forwarding = config.OSX_ENABLE_FWD
+  # enable_firewall = config.OSX_ENABLE_FIREWALL
+  run_cmds(add_permissions, enable_forwarding)
+  print('Permissions set')
+
+def restore_permissions():
+  print(' Restoring to default permissions...')
+  subtract_permissions = config.OSX_SUBTR_BPF_PERMISSIONS
+  disable_forwarding = config.OSX_DISABLE_FWD
+  # disable_firewall = config.OSX_DISABLE_FIREWALL
+  run_cmds(subtract_permissions, disable_forwarding)
+  print('Permissions restored')
